+++ src/cmd/commands_handler.h
@@ -6,9 +6,18 @@
 
 #define CMD_MAX_BUF_SIZE        128
 
+typedef struct sms_command {
+    char phone[20];
+    char command[CMD_MAX_BUF_SIZE];
+} sms_command_t;
+
 void commands_handler_init();
 
-bool send_command_to_parser(const char *command);
+bool send_command_to_parser(sms_command_t);
+
+void commands_set_response(const char*);
+
+void commands_clear_response();
 
 void commands_handler_task(void *argv);